requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.3
django-cleanup==6.0.0
django-faicon==0.1.0
django-material-icon-widget==0.1.2
django-phonenumber-field==7.0.1
gunicorn==20.1.0
idna==3.4
phonenumbers==8.13.1
Pillow==9.3.0
python-dotenv==0.21.0
PyYAML==5.4.1
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
